Small update
changed  how the  "nearNumber" returns values if they are the same.

Also added more to the "/station-get/local" address.

The calculation of the closest station is not working properly when StationData is given values from the weatther station ingame, but it works with dummy numbers sent from a browser via queries